requirements.txt
anyio==4.6.0
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
paramiko==3.5.0
pycparser==2.22
PyNaCl==1.5.0
pyTelegramBotAPI==4.23.0
python-dotenv==1.0.1
python-telegram-bot==21.6
requests==2.32.3
sniffio==1.3.1
urllib3==2.2.3
